Refactor doc
