+++ src/components/HeaderBar/HeaderBurgerMenu/HeaderBurgerMenu.mdx
@@ -197,7 +197,11 @@ code:`
 
 Voici un exemple de menu construit de façon dynamique.
 
-<Canvas of={HeaderBurgerMenuStories.Generated}
+<Canvas
+	of={HeaderBurgerMenuStories.Generated}
+	story={{
+		height: '500px',
+	}}
 	source={{
 language: 'html',
 format: 'dedent',
